rename {C => c}omponents dir
This is one easy mistake away from causing confusing problems on a case-sensitive file system.
