+++ vscode/webviews/components/ChatModelIcon.tsx
