+++ vscode/webviews/components/UserAvatar.module.css
