+++ vscode/webviews/components/UserAvatar.story.tsx
