+++ vscode/webviews/components/UserAvatar.tsx
