+++ vscode/webviews/components/platform/FloatingWidget.module.css
