+++ vscode/webviews/components/platform/FloatingWidget.tsx
