+++ vscode/webviews/components/platform/Popover.module.css
